Fixed migration to create tags out of news_category snippets

+++ db/migrate/20120523133400_move_page_news_category_data_to_tags.rb
@@ -3,17 +3,16 @@ class MovePageNewsCategoryDataToTags < ActiveRecord::Migration
   def self.up
     Snippet.where(:item_type => 'Page', :name => 'news_category').each do |snippet|
       page = snippet.item
-      page.category_list << snippet.text
-      page.save
-      snippet.destroy
+      if page.update_attribute(:category_list, page.category_list + [snippet.text])
+        snippet.destroy
+      end
     end
   end
 
   def self.down
     Page.all.each do |page|
-      page.category_list.each {|category| Snippet.create(:item => page, :name => 'news_category', :text => category.humanize)}
-      page.category_list = []
-      page.save
+      page.category_list.each {|category| Snippet.create(:item => page, :name => 'news_category', :text => category.titleize.humanize)}
+      page.update_attribute(:category_list, [])
     end
   end
